requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
sqlalchemy==2.0.27
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.1
pydantic[email]==2.6.1
